Adding soap communication
