+++ MonolitApp/agent-app/src/main/java/META-INF/JAXB/episode_xjc_2.xjb
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<bindings version="2.1" xmlns="http://java.sun.com/xml/ns/jaxb">
+  <!--
+
+This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
+See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
+Any modifications to this file will be lost upon recompilation of the source schema. 
+Generated on: 2020.06.21 at 12:06:09 AM CEST 
+
+  -->
+  <bindings scd="x-schema::tns" xmlns:tns="http://www.app.agent/gen">
+    <schemaBindings map="false">
+      <package name="agent.app.gen"/>
+    </schemaBindings>
+    <bindings scd="tns:getPublisherRequestsRequest">
+      <class ref="agent.app.gen.GetPublisherRequestsRequest"/>
+    </bindings>
+    <bindings scd="tns:getPublisherRequestsByStatusRequest">
+      <class ref="agent.app.gen.GetPublisherRequestsByStatusRequest"/>
+    </bindings>
+    <bindings scd="tns:getPublisherReportsRequest">
+      <class ref="agent.app.gen.GetPublisherReportsRequest"/>
+    </bindings>
+    <bindings scd="tns:getAdReportsRequest">
+      <class ref="agent.app.gen.GetAdReportsRequest"/>
+    </bindings>
+    <bindings scd="tns:getPublisherRequestsResponse">
+      <class ref="agent.app.gen.GetPublisherRequestsResponse"/>
+    </bindings>
+    <bindings scd="~tns:Request">
+      <class ref="agent.app.gen.Request"/>
+    </bindings>
+    <bindings scd="tns:getPublisherReportsResponse">
+      <class ref="agent.app.gen.GetPublisherReportsResponse"/>
+    </bindings>
+    <bindings scd="~tns:Report">
+      <class ref="agent.app.gen.Report"/>
+    </bindings>
+    <bindings scd="tns:getAdReportsResponse">
+      <class ref="agent.app.gen.GetAdReportsResponse"/>
+    </bindings>
+    <bindings scd="~tns:Ad">
+      <class ref="agent.app.gen.Ad"/>
+    </bindings>
+    <bindings scd="~tns:RequestStatusEnum">
+      <typesafeEnumClass ref="agent.app.gen.RequestStatusEnum"/>
+    </bindings>
+  </bindings>
+</bindings>
+
