+++ MonolitApp/agent-app/src/main/java/META-INF/JAXB/episode_xjc_3.xjb
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<bindings version="2.1" xmlns="http://java.sun.com/xml/ns/jaxb">
+  <!--
+
+This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
+See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
+Any modifications to this file will be lost upon recompilation of the source schema. 
+Generated on: 2020.06.21 at 12:08:56 AM CEST 
+
+  -->
+  <bindings scd="x-schema::tns" xmlns:tns="http://www.app.services/carrequestservice/model">
+    <schemaBindings map="false">
+      <package name="services.app.carrequestservice.model"/>
+    </schemaBindings>
+    <bindings scd="tns:getPublisherRequestsRequest">
+      <class ref="services.app.carrequestservice.model.GetPublisherRequestsRequest"/>
+    </bindings>
+    <bindings scd="tns:getPublisherRequestsByStatusRequest">
+      <class ref="services.app.carrequestservice.model.GetPublisherRequestsByStatusRequest"/>
+    </bindings>
+    <bindings scd="tns:getPublisherReportsRequest">
+      <class ref="services.app.carrequestservice.model.GetPublisherReportsRequest"/>
+    </bindings>
+    <bindings scd="tns:getAdReportsRequest">
+      <class ref="services.app.carrequestservice.model.GetAdReportsRequest"/>
+    </bindings>
+    <bindings scd="tns:getPublisherRequestsResponse">
+      <class ref="services.app.carrequestservice.model.GetPublisherRequestsResponse"/>
+    </bindings>
+    <bindings scd="~tns:Request">
+      <class ref="services.app.carrequestservice.model.Request"/>
+    </bindings>
+    <bindings scd="tns:getPublisherReportsResponse">
+      <class ref="services.app.carrequestservice.model.GetPublisherReportsResponse"/>
+    </bindings>
+    <bindings scd="~tns:Report">
+      <class ref="services.app.carrequestservice.model.Report"/>
+    </bindings>
+    <bindings scd="tns:getAdReportsResponse">
+      <class ref="services.app.carrequestservice.model.GetAdReportsResponse"/>
+    </bindings>
+    <bindings scd="~tns:Ad">
+      <class ref="services.app.carrequestservice.model.Ad"/>
+    </bindings>
+    <bindings scd="~tns:RequestStatusEnum">
+      <typesafeEnumClass ref="services.app.carrequestservice.model.RequestStatusEnum"/>
+    </bindings>
+  </bindings>
+</bindings>
+
