+++ MonolitApp/agent-app/src/main/java/services/app/carrequestservice/model/Ad.java
@@ -2,11 +2,11 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
 // See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.06.20 at 08:05:58 PM CEST 
+// Generated on: 2020.06.21 at 12:08:56 AM CEST 
 //
 
 
-package agent.app.gen;
+package services.app.carrequestservice.model;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
