+++ MonolitApp/agent-app/src/main/java/services/app/carrequestservice/model/GetAdReportsResponse.java
@@ -2,11 +2,11 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
 // See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.06.20 at 08:05:58 PM CEST 
+// Generated on: 2020.06.21 at 12:08:56 AM CEST 
 //
 
 
-package agent.app.gen;
+package services.app.carrequestservice.model;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -26,7 +26,7 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent&gt;
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
  *       &lt;sequence&gt;
- *         &lt;element name="reports" type="{http://www.app.agent/gen}Report" maxOccurs="unbounded" minOccurs="0"/&gt;
+ *         &lt;element name="reports" type="{http://www.app.services/carrequestservice/model}Report" maxOccurs="unbounded" minOccurs="0"/&gt;
  *       &lt;/sequence&gt;
  *     &lt;/restriction&gt;
  *   &lt;/complexContent&gt;
@@ -39,8 +39,8 @@ import javax.xml.bind.annotation.XmlType;
 @XmlType(name = "", propOrder = {
     "reports"
 })
-@XmlRootElement(name = "getPublisherReportsResponse")
-public class GetPublisherReportsResponse {
+@XmlRootElement(name = "getAdReportsResponse")
+public class GetAdReportsResponse {
 
     protected List<Report> reports;
