+++ MonolitApp/agent-app/src/main/java/services/app/carrequestservice/model/GetPublisherRequestsByStatusRequest.java
@@ -2,11 +2,11 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
 // See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.06.20 at 08:05:58 PM CEST 
+// Generated on: 2020.06.21 at 12:08:56 AM CEST 
 //
 
 
-package agent.app.gen;
+package services.app.carrequestservice.model;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -26,6 +26,7 @@ import javax.xml.bind.annotation.XmlType;
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
  *       &lt;sequence&gt;
  *         &lt;element name="publisherUser" type="{http://www.w3.org/2001/XMLSchema}string"/&gt;
+ *         &lt;element name="status" type="{http://www.w3.org/2001/XMLSchema}string"/&gt;
  *       &lt;/sequence&gt;
  *     &lt;/restriction&gt;
  *   &lt;/complexContent&gt;
@@ -36,13 +37,16 @@ import javax.xml.bind.annotation.XmlType;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "", propOrder = {
-    "publisherUser"
+    "publisherUser",
+    "status"
 })
-@XmlRootElement(name = "getPublisherReportsRequest")
-public class GetPublisherReportsRequest {
+@XmlRootElement(name = "getPublisherRequestsByStatusRequest")
+public class GetPublisherRequestsByStatusRequest {
 
     @XmlElement(required = true)
     protected String publisherUser;
+    @XmlElement(required = true)
+    protected String status;
 
     /**
      * Gets the value of the publisherUser property.
@@ -66,6 +70,30 @@ public class GetPublisherReportsRequest {
      */
     public void setPublisherUser(String value) {
         this.publisherUser = value;
+    }
+
+    /**
+     * Gets the value of the status property.
+     * 
+     * @return
+     *     possible object is
+     *     {@link String }
+     *     
+     */
+    public String getStatus() {
+        return status;
+    }
+
+    /**
+     * Sets the value of the status property.
+     * 
+     * @param value
+     *     allowed object is
+     *     {@link String }
+     *     
+     */
+    public void setStatus(String value) {
+        this.status = value;
     }
 
 }
