+++ MonolitApp/agent-app/src/main/java/services/app/carrequestservice/model/GetPublisherRequestsRequest.java
@@ -2,20 +2,17 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
 // See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.06.20 at 08:05:58 PM CEST 
+// Generated on: 2020.06.21 at 12:08:56 AM CEST 
 //
 
 
-package agent.app.gen;
+package services.app.carrequestservice.model;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.XmlSchemaType;
 import javax.xml.bind.annotation.XmlType;
-import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
-import org.w3._2001.xmlschema.Adapter5;
 
 
 /**
@@ -28,7 +25,7 @@ import org.w3._2001.xmlschema.Adapter5;
  *   &lt;complexContent&gt;
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
  *       &lt;sequence&gt;
- *         &lt;element name="adId" type="{http://www.w3.org/2001/XMLSchema}long"/&gt;
+ *         &lt;element name="publisherUser" type="{http://www.w3.org/2001/XMLSchema}string"/&gt;
  *       &lt;/sequence&gt;
  *     &lt;/restriction&gt;
  *   &lt;/complexContent&gt;
@@ -39,38 +36,36 @@ import org.w3._2001.xmlschema.Adapter5;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "", propOrder = {
-    "adId"
+    "publisherUser"
 })
-@XmlRootElement(name = "getAdReportsRequest")
-public class GetAdReportsRequest {
+@XmlRootElement(name = "getPublisherRequestsRequest")
+public class GetPublisherRequestsRequest {
 
-    @XmlElement(required = true, type = String.class)
-    @XmlJavaTypeAdapter(Adapter5 .class)
-    @XmlSchemaType(name = "long")
-    protected Long adId;
+    @XmlElement(required = true)
+    protected String publisherUser;
 
     /**
-     * Gets the value of the adId property.
+     * Gets the value of the publisherUser property.
      * 
      * @return
      *     possible object is
      *     {@link String }
      *     
      */
-    public Long getAdId() {
-        return adId;
+    public String getPublisherUser() {
+        return publisherUser;
     }
 
     /**
-     * Sets the value of the adId property.
+     * Sets the value of the publisherUser property.
      * 
      * @param value
      *     allowed object is
      *     {@link String }
      *     
      */
-    public void setAdId(Long value) {
-        this.adId = value;
+    public void setPublisherUser(String value) {
+        this.publisherUser = value;
     }
 
 }
