+++ MonolitApp/agent-app/src/main/java/services/app/carrequestservice/model/GetPublisherRequestsResponse.java
@@ -2,11 +2,11 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
 // See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.06.20 at 08:05:58 PM CEST 
+// Generated on: 2020.06.21 at 12:08:56 AM CEST 
 //
 
 
-package agent.app.gen;
+package services.app.carrequestservice.model;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -26,7 +26,7 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent&gt;
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
  *       &lt;sequence&gt;
- *         &lt;element name="reports" type="{http://www.app.agent/gen}Report" maxOccurs="unbounded" minOccurs="0"/&gt;
+ *         &lt;element name="requests" type="{http://www.app.services/carrequestservice/model}Request" maxOccurs="unbounded" minOccurs="0"/&gt;
  *       &lt;/sequence&gt;
  *     &lt;/restriction&gt;
  *   &lt;/complexContent&gt;
@@ -37,40 +37,40 @@ import javax.xml.bind.annotation.XmlType;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "", propOrder = {
-    "reports"
+    "requests"
 })
-@XmlRootElement(name = "getAdReportsResponse")
-public class GetAdReportsResponse {
+@XmlRootElement(name = "getPublisherRequestsResponse")
+public class GetPublisherRequestsResponse {
 
-    protected List<Report> reports;
+    protected List<Request> requests;
 
     /**
-     * Gets the value of the reports property.
+     * Gets the value of the requests property.
      * 
      * <p>
      * This accessor method returns a reference to the live list,
      * not a snapshot. Therefore any modification you make to the
      * returned list will be present inside the JAXB object.
-     * This is why there is not a <CODE>set</CODE> method for the reports property.
+     * This is why there is not a <CODE>set</CODE> method for the requests property.
      * 
      * <p>
      * For example, to add a new item, do as follows:
      * <pre>
-     *    getReports().add(newItem);
+     *    getRequests().add(newItem);
      * </pre>
      * 
      * 
      * <p>
      * Objects of the following type(s) are allowed in the list
-     * {@link Report }
+     * {@link Request }
      * 
      * 
      */
-    public List<Report> getReports() {
-        if (reports == null) {
-            reports = new ArrayList<Report>();
+    public List<Request> getRequests() {
+        if (requests == null) {
+            requests = new ArrayList<Request>();
         }
-        return this.reports;
+        return this.requests;
     }
 
 }
