+++ MonolitApp/agent-app/src/main/java/services/app/carrequestservice/model/Request.java
@@ -2,11 +2,11 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
 // See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.06.20 at 08:05:58 PM CEST 
+// Generated on: 2020.06.21 at 12:08:56 AM CEST 
 //
 
 
-package agent.app.gen;
+package services.app.carrequestservice.model;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,6 +16,9 @@ import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlSchemaType;
 import javax.xml.bind.annotation.XmlType;
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
+
+import agent.app.converter.DateAPI;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import org.joda.time.DateTime;
 import org.w3._2001.xmlschema.Adapter1;
 import org.w3._2001.xmlschema.Adapter4;
@@ -33,14 +36,14 @@ import org.w3._2001.xmlschema.Adapter5;
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
  *       &lt;sequence&gt;
  *         &lt;element name="id" type="{http://www.w3.org/2001/XMLSchema}long"/&gt;
- *         &lt;element name="status" type="{http://www.app.agent/gen}RequestStatusEnum"/&gt;
+ *         &lt;element name="status" type="{http://www.app.services/carrequestservice/model}RequestStatusEnum"/&gt;
  *         &lt;element name="submitDate" type="{http://www.w3.org/2001/XMLSchema}dateTime"/&gt;
  *         &lt;element name="endUser" type="{http://www.w3.org/2001/XMLSchema}long"/&gt;
  *         &lt;element name="publisherUser" type="{http://www.w3.org/2001/XMLSchema}long"/&gt;
  *         &lt;element name="bundle" type="{http://www.w3.org/2001/XMLSchema}boolean"/&gt;
  *         &lt;element name="startDate" type="{http://www.w3.org/2001/XMLSchema}dateTime"/&gt;
  *         &lt;element name="endDate" type="{http://www.w3.org/2001/XMLSchema}dateTime"/&gt;
- *         &lt;element name="ads" type="{http://www.app.agent/gen}Ad" maxOccurs="unbounded"/&gt;
+ *         &lt;element name="ads" type="{http://www.app.services/carrequestservice/model}Ad" maxOccurs="unbounded"/&gt;
  *       &lt;/sequence&gt;
  *     &lt;/restriction&gt;
  *   &lt;/complexContent&gt;
@@ -97,6 +100,20 @@ public class Request {
     @XmlElement(required = true)
     protected List<Ad> ads;
 
+    @JsonProperty("submitDate")
+    public String getTheSubmitDate() {
+        return DateAPI.StringDateFromDateTime(submitDate);
+    }
+
+    @JsonProperty("startDate")
+    public String getTheStartDate() {
+        return DateAPI.StringDateFromDateTime(startDate);
+    }
+
+    @JsonProperty("endDate")
+    public String getTheEndDate() {
+        return DateAPI.StringDateFromDateTime(endDate);
+    }
     /**
      * Gets the value of the id property.
      *
