+++ MonolitApp/agent-app/src/main/java/services/app/carrequestservice/model/package-info.java
@@ -2,8 +2,8 @@
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
 // See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2020.06.20 at 08:05:58 PM CEST 
+// Generated on: 2020.06.21 at 12:08:56 AM CEST 
 //
 
-@javax.xml.bind.annotation.XmlSchema(namespace = "http://www.app.agent/gen", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
-package agent.app.gen;
+@javax.xml.bind.annotation.XmlSchema(namespace = "http://www.app.services/carrequestservice/model", elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED)
+package services.app.carrequestservice.model;
